learning-copilot/requirements.txt
streamlit==1.40.0
openai==1.58.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
PyGithub==2.5.0
PyPDF2==3.0.1
pandas==2.2.3
numpy==2.2.1
requests==2.32.3
markdown==3.7
beautifulsoup4==4.12.3
Pillow==11.0.0